Refactor output handling to allow for settings append
Some users may feel more comfortable or security settings may make
it better to simply append the updated settings to the end of the
configuration file. This PR therefore redirects much of the
informational output to stderr so it does not end up in the config
file (when using --quiet, --yes, and --dry-run).

Tests for the printers also added.