Improve quiet mode handling and add tests
We no longer have to pass the 'quiet' flag around as a function
parameter since all the functions that use it are methods of the
tstune.Tuner object and can access it that way.

Previously the quiet mode did not handle no changes as gracefully
as it could have; it still prompted for a yes/no instead of just
assuming success.

Test coverage added to cover these cases