Fix recommendations for parallelism in PG 9.6
PostgreSQL 9.6 does not support the max_parallel_workers option,
so the tool should not recommend adding or changing that setting
if being run on a 9.6 instance.

This requires using pg_config to detect which version of PG is
being run on the machine. Given the output of pg_config, we can
also simplify our look up logic for possible paths where the conf
file might exist.

This commit adds support for coverting PostgreSQL version strings
into different major versions which is then used by the tune tool.