Add line for when database was last tuned
This line can then potentially be used by the main database to
potentially present notices to the user that their settings may be
out of date. Additionally we can use the value for telemetrics.

Additionally, this commit fixes when effective_io_concurrency is
set, i.e., only on Linux systems.