Increase default max_connections
The old default max_connections of 20 was a bit low for some use
cases. Increasing the max_connections to 50 should be more reasonable.
In addition, we add the scaffolding to allow a user to override the
given max_connections recommendation with their own value based on
their usage pattern. Since we are doing this, we need to handle
adjusting the work_mem based on max_connections since with more
connections you will want to reduce the work_mem for each.