Refactor tests for pkg/pgtune for maintainability
The updates to pgtune.MiscSettingsGroup's tests provide a good
template to do for other SettingsGroups for more thorough coverage.
While adding this support, it simplifies the test functions to be
of the same form for most SettingsGroup so it can be furthered
refactored, leaving less bespoke code for each and making the whole
thing more maintainable.

Tests for MemorySettingsGroup will be done in a separate PR as its
test cases are less amenable to the simplified structure for the
others.