Update parallelism settings for TimescaleDB bg workers
With the upcoming support for additional background worker tasks
in TimescaleDB, it is important that timescaledb-tune sets more
appropriate parallelism settings. This commit adds setting the
TimescaleDB config option 'max_background_workers' to a default of
4 (for now) which should give at least a few background workers
for TimescaleDB to work with. This value, along with the number of
CPUs and a minimum number of processes for built-ins, helps create
a more useful number of 'max_processes'. This resolves the problem
where we might not obtain useful parallelism or have TimescaleDB
background workers run if there are not enough CPUs.