Add --restore flag
In the case that our tool ends up giving a configuration the user
does not like, we add a --restore flag to allow the user to replace
their conf file to a state previously backed up and stored in the
temp dir.

In testing this, we also found a bug where if the new conf file is
shorter than the current one, it will leave garbled text at the
end. This bug is fixed by first truncating the file and then
writing the new contents.