Add ability to override the machine's PG version
Like memory and CPU, users may want to specify their version of
PostgreSQL to use for recommendations. For example, when generating
configuration files for another machine that the tool is being
run on. This also obviates the need for pg_config to be installed
on the machine where the tool is being run since it can just be
provided.