Improve recording of timescaledb-tune parameters
Previously tstune.Tuner would always write new lines for parameters
it generated, such as timescaledb.last_tuned_version. This created a
messy end to conf files when tuned multiple times.

This PR updates Tuner to search for its parameters first, and update
them in place if they exist. If not, then it appends them to the end
of the file. Additionally, support is added for removing duplicate
lines (and potentially other transformations, e.g., commenting out
unused lines instead). For now this will clean up the mess we may
have created before with multiple lines for timescaledb.last_tuned
and timescaledb.last_tuned_version. In the future we can leverage
this to warn users about parameters being overridden later in the
file, comment out unused duplicates, and other 'linting' changes.

This also improves the correctness of the regexes we generate for
parameters but properly escaping meta characters.