Increase number of background workers and conns
With the introduction of continuous aggregates in TimescaleDB 1.3,
more background workers will be needed for the database to run
effectively. In addition, it is probably advisable to update the
max_connections parameter in tandem with this change so that the
background workers TimescaleDB spawns don't eat too much into the
connections available.

This PR also fixes a bug where the recommendation for
max_parallel_per_gather was not being returned for PG 9.6 due to a
off-by-1 error. Tests were added to catch this in the future.