WIP, will be rebased: Connect to PostgreSQL
Relying on a single file for determining the PostgreSQL settings is not
enough to know which settings will take effect.
PostgreSQL has configuration includes[1] that are not taken into account
when only reading the main configuration file.

By connecting to PostgreSQL and reading all the relevant settings from
pg_settings, we can avoid duplicating all of this include logic and
leave all the parsing of configuration files to PostgreSQL.

Allowing a connection string to get the source of the configuration also
allows to run this tool on a remote system and write configuration files
locally; this can be very useful if the output of timescaledb-tune is
supposed to be part of configuration management (e.g. Ansible, puppet)

[1] https://www.postgresql.org/docs/current/config-setting.html#CONFIG-INCLUDES