Add autovacuum recommendations
The default max of autovacuum workers and their naptime can be a
bottleneck when a database has many chunks that need to be vacuumed
and analyzed. Also, if some chunks are larger than others, it may
end blocking smaller chunks from getting processed in a timely
fashion as the system is waiting for autovacuum workers to be freed
up.