Add flag for specifying the size of the WAL disk
Previously the values for min_wal_size and max_wal_size could be
too aggressive for machines with shared WAL and data disks or with
small WAL disks. To that end, this PR introduces a new flag for
specifying the size of the WAL disk (or, the amount of disk space
reserved for the WAL) to base calculations on. The max_wal_size
will be up to 80% of the space given. The min_wal_size is set to
half of the max. If the flag is empty, the default WAL values for
PostgreSQL are used.