Allow sizes to be specified without suffix (bytes)
In some environments, for example k8s, we can pass along Memory or Disk
sizes directly, however currently we need to ensure we add a suffix to
these sizes, as the tool requires a unit specifier. The bytes specifier
was unusable, this meant we needed to do conversion arithmetic, which
in some environments even requires shell arithmentic (yikes).

With this change we can pass on bytes values directly without the need
of conversion.

The pg_size_bytes function of PostgreSQL also allows the same behaviour,
defaulting to bytes if no unit is specified.

The previous definition of B (bytes) already resulted in an error
message and was therefore unusable.