Update max_connections calculation
The default of 100 for max_connections can be a bit high for
instances with less memory; work_mem would be smaller and thus
cause more queries to spill to disk than necessary. This change
creates a step function to allow for smaller max_connection values
for memory sizes below 6GB. A user can still override to a higher
amount using the appropriate flag.

Additionally, this PR fixes a problem with user set max_connections
where the user's value would be used for memory calculations, but
would not necessarily be used for the actual max_connections value.
Previously, if the user's max_connections value was smaller than
the default, the memory calculation would use it to determine the
appropriate work_mem, but ultimately the default max_connections
would be set in the config file. This could lead to incorrect
behavior and overcommitment of memory.