Support for tuning with less than 1 CPU
Kubernetes allows configuring compute resources in fractions of a single CPU,
so-called milliCPUs (e.g. 500 milliCPUs = 0.5 CPU). Use milliCPUs internally
for all calculations. Support additional --milli-cpus parameter to specify
less than 1 CPU in milliCPUs. When it is not null, it is applied instead of
--cpus.

We don't do any special tuning for fractions of CPU, basically treating it
as 1 full CPU. However, the tuning we perform based on CPU cores is either
almost not affected by the difference (work_mem), or requires more than 1
CPU (parallel queries), so in practice it should not change any tunable
parameters in a notable way.