Don't require exactly one space before or after =
My postgresql.conf looks like this:

	checkpoint_timeout = 30min   # Max. time between checkpoints; Default: 5min
	max_wal_size       = 4GB     # Default: 1G
	min_wal_size       = 256MB   # Default: 80M

It would fail to detect these keys as the tuneRegexFmt and
tuneRegexQuotedFmt require exactly " = "; this changes that to "\s*=\s*"
to accept any whitespace. It doesn't make any effort to write these
spaces back – might be nice too, but I couldn't be bothered 😅

Also use ` for these strings so you don't need to double up the
backslashes.